Comments with auth, obs with search, etc.
This basically compeltes the hack day goals, but I think it also presents a
decent framework for how to use webpack, redux, and react within our Rails app
and with our API. It's based on the tutorial at http://redux.js.org, so that's
worth a read. It uses the tutorial file structure, splitting components,
containers, reducers, and actions into different directories, which seems
sensible to me. The place to start is webpack-entry.js, where webpack starts
building its list of dependencies.

I like the file structure, and I like the flow of data is unidirectional and
fairly easy to follow. I don't like how many things you have to edit to
execute any change on the application state (make an action type, make an
action, make a reducer).